% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj.R
\name{fgls}
\alias{fgls}
\title{Generalized Least Squares Objective Function}
\usage{
fgls(Sigmatheta, Sigma)
}
\arguments{
\item{Sigmatheta}{Model-implied variance-covariance matrix
(\eqn{\boldsymbol{\Sigma}\left( \boldsymbol{\theta} \right)}).}

\item{Sigma}{Variance-covariance matrix.
(\eqn{\boldsymbol{\Sigma}}).}
}
\description{
Calculates the generalized least squares objective function.
}
\author{
Ivan Jacob Agaloos Pesigan
}
