% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj.R
\name{fml}
\alias{fml}
\title{Maximum Likelihood Objective Function}
\usage{
fml(Sigmatheta, Sigma)
}
\arguments{
\item{Sigmatheta}{Model-implied variance-covariance matrix
(\eqn{\boldsymbol{\Sigma}\left( \boldsymbol{\theta} \right)}).}

\item{Sigma}{Variance-covariance matrix.
(\eqn{\boldsymbol{\Sigma}}).}
}
\description{
Calculates the maximum likelihood objective function.
}
\seealso{
Other objective functions: 
\code{\link{fgls}()},
\code{\link{fuls}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{objective functions}
\keyword{objective}
