% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj.R
\name{gls}
\alias{gls}
\title{Generalized Least Squares Objective Function}
\usage{
gls(Sigmatheta, Sigma)
}
\arguments{
\item{Sigmatheta}{Model-implied variance-covariance matrix
(\eqn{\boldsymbol{\Sigma}\left( \boldsymbol{\theta} \right)}).}

\item{Sigma}{Variance-covariance matrix
(\eqn{\boldsymbol{\Sigma}}).}
}
\description{
Calculates the generalized least squares objective function.
}
\seealso{
Other objective functions: 
\code{\link{ml}()},
\code{\link{uls}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{objective functions}
\keyword{optimization}
