% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medsimpleobj.R
\name{medsimpleobj}
\alias{medsimpleobj}
\title{Simple Mediation - Objective function}
\usage{
medsimpleobj(theta, Sigmahat)
}
\arguments{
\item{theta}{Numeric vector.
Vector of parameters to estimate.
\verb{theta[1] =} \eqn{\alpha},
\verb{theta[2] =} \eqn{\tau^{\prime}},
\verb{theta[3] =} \eqn{\beta},
\verb{theta[4] =} \eqn{\sigma^{2}_{X}},
\verb{theta[5] =} \eqn{\sigma^{2}_{\varepsilon_{M}}},
\verb{theta[6] =} \eqn{\sigma^{2}_{\varepsilon_{Y}}}.}

\item{Sigmahat}{Matrix.
\eqn{3 \times 3} sample variance-covariance matrix.
\verb{Sigmahat[1,1] =} \eqn{\mathrm{Var} \left( X \right)},
\verb{Sigmahat[2,2] =} \eqn{\mathrm{Var} \left( M \right)},
\verb{Sigmahat[3,3] =} \eqn{\mathrm{Var} \left( Y \right)}.}
}
\description{
Simple Mediation - Objective function
}
\examples{
X <- jeksterslabRdatarepo::thirst
Sigmahat <- cov(X)
jeksterslabRdist::opt(
  FUN = medsimpleobj,
  start_values = rep(x = 0.20, times = 6),
  optim = TRUE,
  Sigmahat = Sigmahat
)
}
\author{
Ivan Jacob Agaloos Pesigan
}
