% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj.R
\name{ml}
\alias{ml}
\title{Maximum Likelihood Objective Function}
\usage{
ml(Sigmatheta, Sigma)
}
\arguments{
\item{Sigmatheta}{Model-implied variance-covariance matrix
(\eqn{\boldsymbol{\Sigma}\left( \boldsymbol{\theta} \right)}).}

\item{Sigma}{Variance-covariance matrix
(\eqn{\boldsymbol{\Sigma}}).}
}
\description{
Calculates the maximum likelihood objective function.
}
\seealso{
Other objective functions: 
\code{\link{gls}()},
\code{\link{uls}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{objective functions}
\keyword{optimization}
