% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{ram_e}
\alias{ram_e}
\title{Reticular Action Model - Residuals
\eqn{\left( \hat{\Sigma} - \hat{\Sigma} \left( \hat{\theta} \right) \right)}}
\usage{
ram_e(Sigmahat, Sigmathetahat = NULL, Ahat, Shat, F, I)
}
\arguments{
\item{Sigmahat}{Matrix.
Estimated variance-covariance matrix
\eqn{\left( \hat{\boldsymbol{\Sigma}} \right)}}

\item{Sigmathetahat}{Matrix.
Model-implied variance-covariance matrix
as a function of estimated parameters
\eqn{\left( \boldsymbol{\Sigma}\left( \hat{\boldsymbol{\theta}} \right) \right)}.}

\item{Ahat}{Matrix.
Estimated asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{Shat}{Matrix.
Estimated symmetric paths (double-headed arrows),
representing variances and covariances.}

\item{F}{Matrix.
Filter matrix
used to select the observed variables.}

\item{I}{Matrix.
Identity matrix.}
}
\description{
\deqn{
  \hat{\Sigma}
  -
  \hat{\Sigma}
  \left(
    \hat{\theta}
  \right)
}
}
\seealso{
Other SEM notation functions: 
\code{\link{ram_M}()},
\code{\link{ram_Sigmatheta}()},
\code{\link{ram_S}()},
\code{\link{ram_mutheta}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{SEM notation functions}
\keyword{matrix}
\keyword{ram}
