% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{ram_M}
\alias{ram_M}
\title{Reticular Action Model - Mean Structure Vector
\eqn{\left( \mathbf{M} \right)}}
\usage{
ram_M(A, F, I, mu)
}
\arguments{
\item{A}{Matrix.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{F}{Matrix.
Filter matrix
used to select the observed variables.}

\item{I}{Matrix.
Identity matrix.}

\item{mu}{Numeric vector.
Vector of means
\eqn{\left( \boldsymbol{\mu} \left( \boldsymbol{\theta} \right) \right)}.}
}
\value{
Returns the mean structure vector \eqn{\left( \mathbf{M} \right)}
derived from the \eqn{\mathbf{A}},
\eqn{\mathbf{F}},
\eqn{\mathbf{I}}, matrices and
\eqn{\left( \boldsymbol{\mu} \left( \boldsymbol{\theta} \right) \right)} vector.
}
\description{
Mean structure vector
\eqn{\left( \mathbf{M} \right)}
using the Reticular Action Model notation.
}
\details{
\deqn{\mathbf{M}
  =
  \mathbf{F}
  \left(
    \mathbf{I}
    -
    \mathbf{A}
  \right)^{-1}
  \boldsymbol{\mu}
  \left(
    \boldsymbol{\theta}
  \right)
}
}
\examples{
A <- matrix(
  data = c(
    0, 0.26^(1 / 2), 0,
    0, 0, 0.26^(1 / 2),
    0, 0, 0
  ),
  ncol = 3
)
F <- I <- diag(3)
mu <- c(100, 100, 100)
ram_M(A = A, F = F, I = I, mu = mu)
}
\references{
McArdle, J. J. (2013).
The development of the RAM rules for latent variable structural equation modeling.
In A. Maydeu-Olivares & J. J. McArdle (Eds.),
\emph{Contemporary Psychometrics: A festschrift for Roderick P. McDonald} (pp. 225--273).
Lawrence Erlbaum Associates.

McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology}, \emph{37} (2), 234--251.
}
\seealso{
Other SEM notation functions: 
\code{\link{ram_Sigmatheta}()},
\code{\link{ram_S}()},
\code{\link{ram_e}()},
\code{\link{ram_mutheta}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{SEM notation functions}
\keyword{matrix}
\keyword{ram}
