% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{ram_M_linreg}
\alias{ram_M_linreg}
\title{Reticular Action Model
(Mean Structure \eqn{\mathbf{M}}
from Regression Slopes
and Means of Observed Variables
of a \eqn{k}-Variable Linear Regression Model)}
\usage{
ram_M_linreg(beta_slopes, muy, muX)
}
\arguments{
\item{beta_slopes}{Vector or \code{k - 1} by \code{1} matrix.
\eqn{\left( k - 1 \right) \times 1}
vector of regression slopes.}

\item{muy}{Numeric.
Mean of the regressand variable
\eqn{\mu_{\mathbf{y}}}.}

\item{muX}{Vector or \code{k - 1} by \code{1} matrix.
\eqn{\left( k - 1 \right) \times 1} vector
of means of
the regressors
\eqn{
      {X}_{2},
      {X}_{3},
      \dots,
      {X}_{k}
    }
excluding the mean for the vector of constants
(\eqn{\boldsymbol{\mu}_{\mathbf{X}}}).}
}
\description{
Mean structure column vector
(\eqn{\mathbf{M}})
using the Reticular Action Model notation
from regression slopes
and means of observed variables
of a \eqn{k}-variable Linear Regression Model.
}
\details{
The parameters of a linear regression model
for the mean structure
are
the slopes
(\eqn{\boldsymbol{\beta}_{\mathrm{slopes}}}),
the mean of
the regressand
\eqn{\mu_\mathbf{y}}
and
the means of
the regressors
\eqn{
    {X}_{2},
    {X}_{3},
    \dots,
    {X}_{k}
  }
excluding the mean for the vector of constants
(\eqn{\boldsymbol{\mu}_{\mathbf{X}}}).
The parameters form the following matrices,
\deqn{
  \mathbf{A}_{k \times k}
  =
  \begin{bmatrix}
    0 & \beta_2 & \beta_3 & \cdots & \beta_k \\
    0 & 0       & 0       & \cdots & 0       \\
    0 & 0       & 0       & \cdots & 0       \\
    0 & 0       & 0       & \cdots & 0       \\
    0 & 0       & 0       & \cdots & 0
  \end{bmatrix} ,
}
\deqn{
  \boldsymbol{\mu}_{k \times 1}
  =
  \begin{bmatrix}
    \mu_{\mathbf{y}} \\
    \mu_{X_2} \\
    \mu_{X_3} \\
    \vdots    \\
    \mu_{X_k}
  \end{bmatrix} ,
}
\deqn{
  \mathbf{I}_{k \times k}
  =
  \begin{bmatrix}
    1 & 0   & 0   & \cdots & 0   \\
    0 & 1   & 0   & \cdots & 0   \\
    0 & 0   & 1   & \cdots & 0   \\
    0 & 0   & 0   & 1      & 0   \\
    0 & 0   & 0   & \cdots & 1
  \end{bmatrix} ,
}
and
\deqn{
  \mathbf{F}_{k \times k}
  =
  \begin{bmatrix}
    1 & 0   & 0   & \cdots & 0   \\
    0 & 1   & 0   & \cdots & 0   \\
    0 & 0   & 1   & \cdots & 0   \\
    0 & 0   & 0   & 1      & 0   \\
    0 & 0   & 0   & \cdots & 1
  \end{bmatrix} .
}
The mean structure column vector
(\eqn{\mathbf{M}})
is calculated using
\deqn{
    \mathbf{M}
    =
    \mathbf{F}
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{-1}
    \boldsymbol{\mu} .
  }
}
\author{
Ivan Jacob Agaloos Pesigan
}
