% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{ram_mutheta_linreg}
\alias{ram_mutheta_linreg}
\title{Reticular Action Model
(\eqn{\boldsymbol{\mu} \left( \boldsymbol{\theta} \right)}
from Parameters of a \eqn{k}-Variable Linear Regression Model)}
\usage{
ram_mutheta_linreg(beta, muX)
}
\arguments{
\item{beta}{Vector or \code{k} by \code{1} matrix.
The vector
\eqn{\boldsymbol{\beta}}
is a \eqn{k \times 1} vector
of \eqn{k} regression coefficients.}

\item{muX}{Vector or \code{k - 1} by \code{1} matrix.
\eqn{\left( k - 1 \right) \times 1} vector
of means of
the regressors
\eqn{
      {X}_{2},
      {X}_{3},
      \dots,
      {X}_{k}
    }
excluding the mean for the vector of constants
(\eqn{\boldsymbol{\mu}_{\mathbf{X}}}).}
}
\description{
Model-implied mean vector
(\eqn{\boldsymbol{\mu} \left( \boldsymbol{\theta} \right)})
using the Reticular Action Model notation
from parameters of a \eqn{k}-variable linear regression model.
}
\details{
The parameters of a linear regression model
for the mean structure
are
the regression coefficients
(\eqn{\boldsymbol{\beta}}),
and
the means of
the regressors
\eqn{
    {X}_{2},
    {X}_{3},
    \dots,
    {X}_{k}
  }
excluding the mean for the vector of constants
(\eqn{\boldsymbol{\mu}_{\mathbf{X}}}).
The parameters form the following matrices,
\deqn{
  \mathbf{A}_{k \times k}
  =
  \begin{bmatrix}
    0 & \beta_2 & \beta_3 & \cdots & \beta_k \\
    0 & 0       & 0       & \cdots & 0       \\
    0 & 0       & 0       & \cdots & 0       \\
    0 & 0       & 0       & \cdots & 0       \\
    0 & 0       & 0       & \cdots & 0
  \end{bmatrix} ,
}
\deqn{
  \mathbf{M}_{k \times 1}
  =
  \begin{bmatrix}
    \beta_1   \\
    \mu_{X_2} \\
    \mu_{X_3} \\
    \vdots    \\
    \mu_{X_k}
  \end{bmatrix} ,
}
\deqn{
  \mathbf{I}_{k \times k}
  =
  \begin{bmatrix}
    1 & 0   & 0   & \cdots & 0   \\
    0 & 1   & 0   & \cdots & 0   \\
    0 & 0   & 1   & \cdots & 0   \\
    0 & 0   & 0   & 1      & 0   \\
    0 & 0   & 0   & \cdots & 1
  \end{bmatrix} ,
}
and
\deqn{
  \mathbf{F}_{k \times k}
  =
  \begin{bmatrix}
    1 & 0   & 0   & \cdots & 0   \\
    0 & 1   & 0   & \cdots & 0   \\
    0 & 0   & 1   & \cdots & 0   \\
    0 & 0   & 0   & 1      & 0   \\
    0 & 0   & 0   & \cdots & 1
  \end{bmatrix} .
}
The model-implied mean vector
(\eqn{
  \boldsymbol{\mu}
  \left(
    \boldsymbol{\theta}
  \right)
})
is calculated using
\deqn{
    \boldsymbol{\mu}
    \left(
      \boldsymbol{\theta}
    \right)
    =
    \mathbf{F}
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{-1}
    \mathbf{M} .
  }
}
\author{
Ivan Jacob Agaloos Pesigan
}
