% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{ram_residuals}
\alias{ram_residuals}
\title{Reticular Action Model - Residuals
\eqn{ \boldsymbol{ \hat{\Sigma} }
  -
  \boldsymbol{
    \hat{\Sigma}
  }
  \left(
    \boldsymbol{
      \hat{\theta}
    }
  \right)
}}
\usage{
ram_residuals(Sigmahat, Sigmathetahat = NULL, Ahat, Shat, filter)
}
\arguments{
\item{Sigmahat}{Matrix.
Estimated variance-covariance matrix
\eqn{
  \boldsymbol{
    \hat{\Sigma}
  }
}}

\item{Sigmathetahat}{Matrix.
Model-implied variance-covariance matrix
as a function of estimated parameters
\eqn{
  \boldsymbol{
    \hat{\Sigma}
  }
  \left(
    \boldsymbol{
      \hat{\theta}
    }
  \right)
} .}

\item{Ahat}{\eqn{mathbf{\hat{A}}} matrix.
Estimated asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings}

\item{Shat}{\eqn{mathbf{\hat{S}}} matrix.
Estimated symmetric paths (double-headed arrows),
representing variances and covariances.}

\item{filter}{\verb{k x p} numeric matrix
\eqn{\mathbf{F}_{k \times p}}.
Filter matrix used to select variables.}
}
\description{
\deqn{
  \boldsymbol{
    \hat{\Sigma}
  }
  -
  \boldsymbol{
    \hat{\Sigma}
  }
  \left(
    \boldsymbol{
      \hat{\theta}
    }
  \right)
}
where
\deqn{
  \boldsymbol{
    \hat{\Sigma}
  }
  \left(
    \hat{
      \boldsymbol{\theta}
    }
  \right)
  =
  \mathbf{F}
  \left(
    \mathbf{I} - \mathbf{\hat{A}}
  \right)^{-1}
  \mathbf{\hat{S}}
  \left[
    \left(
      \mathbf{I} - \mathbf{\hat{A}}
    \right)^{-1}
  \right]^{\prime}
  \mathbf{F}^{\prime} .
}
Items with a hat (^) are estimates using the sample data.
}
\seealso{
Other SEM notation functions: 
\code{\link{ram_M}()},
\code{\link{ram_Sigmatheta}()},
\code{\link{ram_S}()},
\code{\link{ram_mutheta}()},
\code{\link{ram_sigma2}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{SEM notation functions}
\keyword{matrix}
\keyword{ram}
