% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{ram_S}
\alias{ram_S}
\title{Reticular Action Model - The \eqn{ \mathbf{S} } Matrix}
\usage{
ram_S(Sigmatheta, A, filter)
}
\arguments{
\item{Sigmatheta}{Matrix.
Model-implied variance-covariance matrix
\eqn{ \boldsymbol{ \Sigma } \left( \boldsymbol{ \theta } \right) } .}

\item{A}{\verb{p x p} numeric matrix
\eqn{\mathbf{A}_{p \times p}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{filter}{\verb{k x p} numeric matrix
\eqn{\mathbf{F}_{k \times p}}.
Filter matrix used to select variables.}
}
\description{
Derives the \eqn{ \mathbf{S} } matrix
using the Reticular Action Model (RAM) notation.
}
\details{
The \eqn{ \mathbf{S} } matrix
is derived using
the \eqn{\mathbf{A}}
matrix
and the model-implied variance-covariance matrix
\eqn{ \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)}
}
\seealso{
Other SEM notation functions: 
\code{\link{ram_M}()},
\code{\link{ram_Sigmatheta}()},
\code{\link{ram_mutheta}()},
\code{\link{ram_res}()},
\code{\link{ram_sigma2}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{SEM notation functions}
\keyword{matrix}
\keyword{ram}
