% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{ram_Sigmatheta_linreg}
\alias{ram_Sigmatheta_linreg}
\title{Reticular Action Model
(\eqn{\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)}
from Parameters of a \eqn{k}-Variable Linear Regression Model)}
\usage{
ram_Sigmatheta_linreg(beta_slopes, sigma2, SigmaX)
}
\arguments{
\item{beta_slopes}{Vector or \code{k - 1} by \code{1} matrix.
\eqn{\left( k - 1 \right) \times 1}
vector of regression slopes.}

\item{sigma2}{Numeric.
Variance of the error term
\eqn{\boldsymbol{\varepsilon}}
(\eqn{\sigma^2}).}

\item{SigmaX}{Matrix.
Covariances of
\eqn{
    {X}_{2},
    {X}_{3},
    \dots,
    {X}_{k}
  }
(\eqn{\boldsymbol{\Sigma}_{\mathbf{X}}}).}
}
\value{
Returns the model-implied
variance-covariance matrix
(\eqn{\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)}).
}
\description{
Model-implied variance-covariance matrix
(\eqn{\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)})
using the Reticular Action Model notation
from parameters of a \eqn{k}-variable linear regression model.
}
\details{
The parameters of a linear regression model
for the covariance structure
are
the slopes
(\eqn{\boldsymbol{\beta}_{\mathrm{slopes}}}),
the variance of the error term \eqn{\boldsymbol{\varepsilon}}
(\eqn{\sigma^2}),
and
the covariances of
\eqn{
    {X}_{2},
    {X}_{3},
    \dots,
    {X}_{k}
  }
(\eqn{\boldsymbol{\Sigma}_{\mathbf{X}}}).
The parameters form the following matrices,
\deqn{
  \mathbf{A}_{k \times k}
  =
  \begin{bmatrix}
    0 & \beta_2 & \beta_3 & \cdots & \beta_k \\
    0 & 0       & 0       & \cdots & 0       \\
    0 & 0       & 0       & \cdots & 0       \\
    0 & 0       & 0       & \cdots & 0       \\
    0 & 0       & 0       & \cdots & 0
  \end{bmatrix} ,
}
\deqn{
  \mathbf{S}_{k \times k}
  =
  \begin{bmatrix}
    \sigma^2 & 0                         & 0                         & \cdots & 0   \\
    0        & \mathrm{Var}_{X_2}        & 0                         & \cdots & 0   \\
    0        & \mathrm{Cov}_{X_{3}X_{2}} & \mathrm{Var}_{X_3}        & \cdots & 0   \\
    0        & \vdots                    & \vdots                    & \ddots & 0   \\
    0        & \mathrm{Cov}_{X_{k}X_{2}} & \mathrm{Cov}_{X_{k}X_{3}} & \cdots & \mathrm{Var}_{X_k}
  \end{bmatrix} ,
}
\deqn{
  \mathbf{I}_{k \times k}
  =
  \begin{bmatrix}
    1 & 0   & 0   & \cdots & 0   \\
    0 & 1   & 0   & \cdots & 0   \\
    0 & 0   & 1   & \cdots & 0   \\
    0 & 0   & 0   & 1      & 0   \\
    0 & 0   & 0   & \cdots & 1
  \end{bmatrix} ,
}
and
\deqn{
  \mathbf{F}_{k \times k}
  =
  \begin{bmatrix}
    1 & 0   & 0   & \cdots & 0   \\
    0 & 1   & 0   & \cdots & 0   \\
    0 & 0   & 1   & \cdots & 0   \\
    0 & 0   & 0   & 1      & 0   \\
    0 & 0   & 0   & \cdots & 1
  \end{bmatrix} .
}
The model-implied variance-covariance matrix
(\eqn{
  \boldsymbol{\Sigma}
  \left(
    \boldsymbol{\theta}
  \right)
})
is calculated using
\deqn{
    \boldsymbol{\Sigma}
    \left(
      \boldsymbol{\theta}
    \right)
    =
    \mathbf{F}
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{-1}
    \mathbf{S}
    \left[
      \left(
        \mathbf{I} - \mathbf{A}
      \right)^{-1}
    \right]^{\prime}
    \mathbf{F}^{\prime} .
  }
}
\author{
Ivan Jacob Agaloos Pesigan
}
