% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{ramM}
\alias{ramM}
\title{Reticular Action Model - Mean Structure Vector \eqn{\mathbf{M}}}
\usage{
ramM(mu, A, filter)
}
\arguments{
\item{mu}{\verb{k x 1} numeric vector
\eqn{\boldsymbol{\mu} \left( \boldsymbol{\theta} \right)_{k \times 1}} .
Model-implied meam vector.}

\item{A}{\verb{m x m} numeric matrix
\eqn{\mathbf{A}_{m \times m}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{filter}{\verb{k x m} numeric matrix
\eqn{\mathbf{F}_{k \times m}}.
Filter matrix used to select variables.}
}
\value{
Returns the mean structure vector \eqn{\mathbf{M}}
derived from the \eqn{\mathbf{A}}, \eqn{\mathbf{F}},
\eqn{ \mathbf{I}}, matrices and
\eqn{ \boldsymbol{\mu} \left( \boldsymbol{\theta} \right)}
vector.
}
\description{
Derives the mean structure vector \eqn{\mathbf{M}}
using the Reticular Action Model (RAM) notation.
}
\details{
The mean structure vector \eqn{\mathbf{M}}
as a function of Reticular Action Model (RAM) matrices is given by

\deqn{
    \mathbf{M}
    =
    \left( \mathbf{I} - \mathbf{A} \right)^{-1}
    \mathbf{F}^{T} \boldsymbol{\mu} \left( \boldsymbol{\theta} \right)
  }

where
\itemize{
\item \eqn{\mathbf{A}_{m \times m}} represents asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings,
\item \eqn{\mathbf{I}_{m \times m}} represents an identity matrix,
\item \eqn{\mathbf{F}_{k \times m}} represents the filter matrix
used to select the observed variables,
\item \eqn{\boldsymbol{\mu} \left( \boldsymbol{\theta} \right)}
is the \eqn{k \times 1} model-implied mean vector
\item \eqn{k} number of observed variables,
\item \eqn{q} number of latent variables, and
\item \eqn{m} number of observed and latent variables, that is \eqn{k + q} .
}
}
\references{
McArdle, J. J. (2013).
The development of the RAM rules for latent variable structural equation modeling.
In A. Maydeu-Olivares & J. J. McArdle (Eds.),
\emph{Contemporary Psychometrics: A festschrift for Roderick P. McDonald} (pp. 225--273).
Lawrence Erlbaum Associates.

McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology}, \emph{37} (2), 234--251.
}
\seealso{
Other SEM notation functions: 
\code{\link{ramSigmatheta}()},
\code{\link{rammutheta}()},
\code{\link{ramsigma2}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{SEM notation functions}
\keyword{matrix}
\keyword{ram}
