% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{ramres}
\alias{ramres}
\title{Reticular Action Model - Residual Matrix
\eqn{
  \boldsymbol{\hat{\Sigma}} - \boldsymbol{\hat{\Sigma}}
  \left(\boldsymbol{\hat{\theta}}\right)
  }}
\usage{
ramres(Sigmahat, Sigmathetahat = NULL, Ahat, Shat, filter)
}
\arguments{
\item{Sigmahat}{Matrix.
Estimated variance-covariance matrix
\eqn{
  \boldsymbol{
    \hat{\Sigma}
  }
}}

\item{Sigmathetahat}{Matrix.
Model-implied variance-covariance matrix
as a function of estimated parameters
\eqn{
  \boldsymbol{
    \hat{\Sigma}
  }
  \left(
    \boldsymbol{
      \hat{\theta}
    }
  \right)
} .}

\item{Ahat}{\eqn{mathbf{\hat{A}}} matrix.
Estimated asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings}

\item{Shat}{\eqn{mathbf{\hat{S}}} matrix.
Estimated symmetric paths (double-headed arrows),
representing variances and covariances.}

\item{filter}{\verb{k x p} numeric matrix
\eqn{\mathbf{F}_{k \times p}}.
Filter matrix used to select variables.}
}
\description{
The residual matrix, that is, the discrepancy
between the fitted covariance matrix and the fitted model-implied covariance matrix
is given by
\deqn{
  \boldsymbol{\hat{\Sigma}} - \boldsymbol{\hat{\Sigma}}
  \left(\boldsymbol{\hat{\theta}}\right)
  }
where
\deqn{
    \boldsymbol{\hat{\Sigma}} \left(\boldsymbol{\hat{\theta}} \right)
  =
  \mathbf{F}
  \left(
    \mathbf{I} - \mathbf{\hat{A}}
  \right)^{-1}
  \mathbf{\hat{S}}
  \left[
    \left(
      \mathbf{I} - \mathbf{\hat{A}}
    \right)^{-1}
  \right]^{T}
  \mathbf{F}^{T} .
}
Items with a hat (^) are estimates using the sample data.
}
\seealso{
Other SEM notation functions: 
\code{\link{ramM}()},
\code{\link{ramSigmatheta}()},
\code{\link{rammutheta}()},
\code{\link{ramsigma2}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{SEM notation functions}
\keyword{matrix}
\keyword{ram}
