% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{ramsigma2}
\alias{ramsigma2}
\title{Reticular Action Model - The \eqn{\mathbf{S}} Matrix from \eqn{\sigma^2}}
\usage{
ramsigma2(sigma2, A, start = TRUE)
}
\arguments{
\item{sigma2}{Numeric vector.
Vector of variances \eqn{\sigma^2}.}

\item{A}{\verb{m x m} numeric matrix
\eqn{\mathbf{A}_{m \times m}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{start}{Logical.
If \code{TRUE}, an exogenous variable is positioned as the first element in the matrices.
If \code{FALSE}, an exogenous variable is positioned as the last element in the matrices.}
}
\value{
Returns the  \eqn{\mathbf{S}} matrix.
}
\description{
Derives the \eqn{\mathbf{S}} matrix
using the Reticular Action Model (RAM) notation
from variable variances \eqn{\sigma^2}.
The off-diagonal elements of the \eqn{\mathbf{S}} matrix are assumed to be zeroes.
}
\details{
The \eqn{\mathbf{S}} matrix is derived using the \eqn{\mathbf{A}} matrix
and sigma squared \eqn{\left( \sigma^2 \right)} vector (variances).
\strong{Note that the first or last (see \code{start} argument) element
in the \code{A} and \code{S} matrices should be an exogenous variable.}
}
\references{
McArdle, J. J. (2013).
The development of the RAM rules for latent variable structural equation modeling.
In A. Maydeu-Olivares & J. J. McArdle (Eds.),
\emph{Contemporary Psychometrics: A festschrift for Roderick P. McDonald} (pp. 225--273).
Lawrence Erlbaum Associates.

McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology}, \emph{37} (2), 234--251.
}
\seealso{
Other SEM notation functions: 
\code{\link{ramM}()},
\code{\link{ramSigmatheta}()},
\code{\link{rammutheta}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{SEM notation functions}
\keyword{matrix}
\keyword{ram}
