% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj.R
\name{uls}
\alias{uls}
\title{Unweighted Least Squares Objective Function}
\usage{
uls(Sigmatheta, Sigma)
}
\arguments{
\item{Sigmatheta}{Model-implied variance-covariance matrix
(\eqn{\boldsymbol{\Sigma}\left( \boldsymbol{\theta} \right)}).}

\item{Sigma}{Variance-covariance matrix
(\eqn{\boldsymbol{\Sigma}}).}
}
\description{
Calculates the unweighted least squares objective function.
}
\seealso{
Other objective functions: 
\code{\link{gls}()},
\code{\link{ml}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{objective functions}
\keyword{optimization}
