% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{ram_e}
\alias{ram_e}
\title{Reticular Action Model (Residuals)}
\usage{
ram_e(Sigmahat, Sigmathetahat = NULL, Ahat, Shat, F, I)
}
\arguments{
\item{Sigmahat}{Matrix.
Estimated variance-covariance matrix
(\eqn{\hat{\boldsymbol{\Sigma}}})}

\item{Sigmathetahat}{Matrix.
Model-implied variance-covariance matrix
as a function of estimated parameters
(\eqn{\boldsymbol{\Sigma}\left( \hat{\boldsymbol{\theta}} \right)}).}

\item{Ahat}{Estimated asymmetric paths,
such as regression coefficients and factor loadings.}

\item{Shat}{Estimated symmetric matrix
representing variances and covariances.}

\item{F}{Filter matrix
used to select the observed variables.}

\item{I}{Identity matrix.}
}
\description{
Reticular Action Model (Residuals)
}
