% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{ram_M}
\alias{ram_M}
\title{Reticular Action Model - Mean Structure Vector
\eqn{
  \mathbf{M}
}}
\usage{
ram_M(A, F, I, mu, filter = TRUE)
}
\arguments{
\item{A}{\eqn{\mathbf{A}} matrix.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{F}{\eqn{\mathbf{F}} matrix.
Filter matrix
used to select the observed variables.}

\item{I}{\eqn{\mathbf{I}} matrix.
Identity matrix.}

\item{mu}{Numeric vector.
Vector of means
\eqn{
  \boldsymbol{\mu}
  \left(
    \boldsymbol{\theta}
  \right)
  %(\#eq:sem-mutheta)
} .}

\item{filter}{Logical.
If \code{TRUE}, the \code{F} matrix is used.
If \code{FALSE}, the \code{F} matrix is omitted and the argument \code{F} is ignored.
See Value and Details below for more information.}
}
\value{
Returns the mean structure vector
\eqn{
  \mathbf{M}
}
derived from the
\eqn{\mathbf{A}},
\eqn{\mathbf{F}},
\eqn{\mathbf{I}},
matrices and
\eqn{
  \boldsymbol{\mu}
  \left(
    \boldsymbol{\theta}
  \right)
  %(\#eq:sem-mutheta)
}
vector.
}
\description{
Derives the mean structure vector
\eqn{
  \mathbf{M}
}
using the Reticular Action Model (RAM) notation.
}
\details{
The mean structure vector
\eqn{
  \mathbf{M}
}
as a function of Reticular Action Model (RAM) matrices
is given by
\deqn{
  \mathbf{M}
  =
  \mathbf{F}
  \left(
    \mathbf{I} - \mathbf{A}
  \right)^{-1}
  \boldsymbol{\mu}
  \left(
    \boldsymbol{\theta}
  \right)
  %(\#eq:sem-ram-M)
}
}
\examples{
A <- matrix(
  data = c(
    0, 0.26^(1 / 2), 0,
    0, 0, 0.26^(1 / 2),
    0, 0, 0
  ),
  ncol = 3
)
F <- I <- diag(3)
mu <- c(100, 100, 100)
ram_M(A = A, F = F, I = I, mu = mu)
}
\references{
McArdle, J. J. (2013).
The development of the RAM rules for latent variable structural equation modeling.
In A. Maydeu-Olivares & J. J. McArdle (Eds.),
\emph{Contemporary Psychometrics: A festschrift for Roderick P. McDonald} (pp. 225--273).
Lawrence Erlbaum Associates.

McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology}, \emph{37} (2), 234--251.
}
\seealso{
Other SEM notation functions: 
\code{\link{ram_Sigmatheta}()},
\code{\link{ram_S}()},
\code{\link{ram_mutheta}()},
\code{\link{ram_residuals}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{SEM notation functions}
\keyword{matrix}
\keyword{ram}
