% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{ram_mutheta}
\alias{ram_mutheta}
\title{Reticular Action Model (\eqn{\boldsymbol{\mu}})}
\usage{
ram_mutheta(A, F, I, M)
}
\arguments{
\item{A}{Asymmetric paths,
such as regression coefficients and factor loadings.}

\item{F}{Filter matrix
used to select the observed variables.}

\item{I}{Identity matrix.}

\item{M}{Vector of means and intercepts.}
}
\value{
Returns the expected values (\eqn{\boldsymbol{\mu}})
derived from the \eqn{\mathbf{A}},
\eqn{\mathbf{F}},
\eqn{\mathbf{I}}, matrices and
\eqn{\mathbf{M}} vector.
}
\description{
Model-implied \eqn{\boldsymbol{\mu}} vector
using the Reticular Action Model notation.
}
\details{
\deqn{\boldsymbol{\mu} \left( \boldsymbol{\theta} \right)
  =
  \mathbf{F}
  \left( \mathbf{I} - \mathbf{A} \right)^{-1}
  \mathbf{M}
}
}
\examples{
A <- matrix(
  data = c(
    0, 0.26^(1 / 2), 0,
    0, 0, 0.26^(1 / 2),
    0, 0, 0
  ),
  ncol = 3
)
F <- I <- diag(3)
M <- c(100, 49.0098049, 49.0098049)
ram_mutheta(A = A, F = F, I = I, M = M)
}
\references{
McArdle, J. J. (2013).
The development of the RAM rules for latent variable structural equation modeling.
In A. Maydeu-Olivares & J. J. McArdle (Eds.),
\emph{Contemporary Psychometrics: A festschrift for Roderick P. McDonald} (pp. 225--273).
Lawrence Erlbaum Associates.

McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology, 37}(2), 234--251.
}
\seealso{
Other SEM notation functions: 
\code{\link{ram_M}()},
\code{\link{ram_Sigmatheta}()},
\code{\link{ram_S}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{SEM notation functions}
\keyword{matrix}
\keyword{ram}
