% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{ram_mutheta}
\alias{ram_mutheta}
\title{Reticular Action Model - Model-Implied Mean Vector
\eqn{
  \boldsymbol{\mu}
  \left(
    \boldsymbol{\theta}
  \right)
  %(\#eq:sem-mutheta)
}}
\usage{
ram_mutheta(A, F, I, M)
}
\arguments{
\item{A}{\eqn{\mathbf{A}} matrix.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{F}{\eqn{\mathbf{F}} matrix.
Filter matrix
used to select the observed variables.}

\item{I}{\eqn{\mathbf{I}} matrix.
Identity matrix.}

\item{M}{Numeric vector.
Mean structure.
Vector of means and intercepts.}
}
\value{
Returns the means
\eqn{
  \boldsymbol{\mu}
  \left(
    \boldsymbol{\theta}
  \right)
  %(\#eq:sem-mutheta)
}
derived from the
\eqn{\mathbf{A}},
\eqn{\mathbf{F}},
\eqn{\mathbf{I}},
matrices and
\eqn{\mathbf{M}} vector.
}
\description{
Derives the model-implied mean vector
\eqn{
  \boldsymbol{\mu}
  \left(
    \boldsymbol{\theta}
  \right)
  %(\#eq:sem-mutheta)
}
using the Reticular Action Model (RAM) notation.
}
\details{
The model-implied mean vector
\eqn{
  \boldsymbol{\mu}
  \left(
    \boldsymbol{\theta}
  \right)
  %(\#eq:sem-mutheta)
}
as a function of Reticular Action Model (RAM) matrices
is given by
\deqn{
  \boldsymbol{\mu}
  \left(
    \boldsymbol{\theta}
  \right)
  =
  \mathbf{F}
  \left(
    \mathbf{I}
    -
    \mathbf{A}
  \right)^{-1}
  \mathbf{M}
  %(\#eq:sem-ram-mutheta)
}
}
\examples{
A <- matrix(
  data = c(
    0, 0.26^(1 / 2), 0,
    0, 0, 0.26^(1 / 2),
    0, 0, 0
  ),
  ncol = 3
)
F <- I <- diag(3)
M <- c(100, 49.0098049, 49.0098049)
ram_mutheta(A = A, F = F, I = I, M = M)
}
\references{
McArdle, J. J. (2013).
The development of the RAM rules for latent variable structural equation modeling.
In A. Maydeu-Olivares & J. J. McArdle (Eds.),
\emph{Contemporary Psychometrics: A festschrift for Roderick P. McDonald} (pp. 225--273).
Lawrence Erlbaum Associates.

McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology}, \emph{37} (2), 234--251.
}
\seealso{
Other SEM notation functions: 
\code{\link{ram_M}()},
\code{\link{ram_Sigmatheta}()},
\code{\link{ram_S}()},
\code{\link{ram_residuals}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{SEM notation functions}
\keyword{matrix}
\keyword{ram}
