% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{ram_S}
\alias{ram_S}
\title{Reticular Action Model -
The
\eqn{
  \mathbf{S}
}
Matrix
and
The Model-Implied Variance Covariance Matrix
\eqn{
  \boldsymbol{\Sigma}
  \left(
    \boldsymbol{\theta}
  \right)
  %(\#eq:sem-Sigmatheta)
}}
\usage{
ram_S(A, sigma2, F, I, both = FALSE, SigmaMatrix = TRUE, filter = TRUE)
}
\arguments{
\item{A}{\eqn{\mathbf{A}} matrix.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{sigma2}{Numeric vector.
Vector of variances \eqn{\sigma^2}.
\strong{The first element should be the variance of an exogenous variable.}}

\item{F}{\eqn{\mathbf{F}} matrix.
Filter matrix
used to select the observed variables.}

\item{I}{\eqn{\mathbf{I}} matrix.
Identity matrix.}

\item{both}{Logical.
If \code{TRUE}, returns both
the
\eqn{
  \mathbf{S}
}
marix
and
the model-implied variance-covariance matrix
\eqn{
  \boldsymbol{\Sigma}
  \left(
    \boldsymbol{\theta}
  \right)
  %(\#eq:sem-Sigmatheta)
}
in a list.}

\item{SigmaMatrix}{Logical.
If \code{TRUE},
returns
the model-implied variance-covariance matrix
\eqn{
  \boldsymbol{\Sigma}
  \left(
    \boldsymbol{\theta}
  \right)
  %(\#eq:sem-Sigmatheta)
} .
If \code{FALSE},
returns the
\eqn{
  \mathbf{S}
}
matrix.
Ignored if \code{both = TRUE}.}

\item{filter}{Logical.
If \code{TRUE}, the \code{F} matrix is used.
If \code{FALSE}, the \code{F} matrix is omitted and the argument \code{F} is ignored.
See Value and Details below for more information.}
}
\value{
If \code{both = TRUE},
returns both
the
\eqn{
  \mathbf{S}
}
matrix
and
the model-implied variance-covariance matrix
\eqn{
  \boldsymbol{\Sigma}
  \left(
    \boldsymbol{\theta}
  \right)
  %(\#eq:sem-Sigmatheta)
}
derived from the
\eqn{\mathbf{A}} matrix and
\eqn{\sigma^2}
vector.
in a list.
Otherwise,
If \code{both = FALSE},
the model-implied variance-covariance matrix
\eqn{
  \boldsymbol{\Sigma}
  \left(
    \boldsymbol{\theta}
  \right)
  %(\#eq:sem-Sigmatheta)
}
is returned if \code{SigmaMatrix = TRUE}.
If \code{both = FALSE},
\eqn{
  \mathbf{S}
}
matrix
is returned if \code{SigmaMatrix = FALSE}.

NOTE:
SigmaMatrix - option to filter out latent variables
If \code{filter = TRUE}, variance-covariance matrix of observed variables.
If \code{filter = FALSE}, combined variance-covariance matrix of observed and latent variables.
S - always the full S matrix with the same dimensions as the input A matrix
}
\description{
Derives the
\eqn{
  \mathbf{S}
}
matrix
and
the model-implied variance-covariance matrix
\eqn{
  \boldsymbol{\Sigma}
  \left(
    \boldsymbol{\theta}
  \right)
  %(\#eq:sem-Sigmatheta)
}
using the Reticular Action Model (RAM) notation.
}
\details{
The
\eqn{
  \mathbf{S}
}
matrix
and
the model-implied variance-covariance matrix
\eqn{
  \boldsymbol{\Sigma}
  \left(
    \boldsymbol{\theta}
  \right)
  %(\#eq:sem-Sigmatheta)
}
are derived using
the \eqn{\mathbf{A}}
matrix
and sigma squared
\eqn{\left( \sigma^2 \right)}
vector (variances).
\strong{Note that the first element in the A ans S matrices should be an exogenous variable.}
}
\examples{
A <- matrix(
  data = c(
    0, 0.26^(1 / 2), 0,
    0, 0, 0.26^(1 / 2),
    0, 0, 0
  ),
  ncol = 3
)
sigma2 <- c(15^2, 15^2, 15^2)
F <- I <- diag(3)
# Returns the model-implied variance-covariance matrix
ram_S(A = A, sigma2 = sigma2, F = F, I = I, SigmaMatrix = TRUE)
# Returns the model-implied S matrix
ram_S(A = A, sigma2 = sigma2, F = F, I = I, SigmaMatrix = FALSE)
# One-factor CFA model--------------------------------------------------------
A <- matrix(data = 0, ncol = 6, nrow = 6)
for (i in 2:6) {
  A[i, 1] <- 0.5
}
sigma2 <- c(1, 1, 1, 1, 1, 1)
F <- I <- diag(nrow(A))
F <- diag(nrow(A) - 1)
F <- cbind(0, F)
ram_S(
  A = A,
  sigma2 = sigma2,
  F = F,
  I = I
)
ram_S(
  A = A,
  sigma2 = sigma2,
  F = F,
  I = I,
  SigmaMatrix = FALSE
)
}
\references{
McArdle, J. J. (2013).
The development of the RAM rules for latent variable structural equation modeling.
In A. Maydeu-Olivares & J. J. McArdle (Eds.),
\emph{Contemporary Psychometrics: A festschrift for Roderick P. McDonald} (pp. 225--273).
Lawrence Erlbaum Associates.

McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology}, \emph{37} (2), 234--251.
}
\seealso{
Other SEM notation functions: 
\code{\link{ram_M}()},
\code{\link{ram_Sigmatheta}()},
\code{\link{ram_mutheta}()},
\code{\link{ram_residuals}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{SEM notation functions}
\keyword{matrix}
\keyword{ram}
