% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{ram_S}
\alias{ram_S}
\title{Reticular Action Model (\eqn{\boldsymbol{\Sigma}} or \eqn{\mathbf{S}})}
\usage{
ram_S(A, sigma2, F, I, SigmaMatrix = TRUE)
}
\arguments{
\item{A}{Asymmetric paths,
such as regression coefficients and factor loadings.}

\item{sigma2}{Vector of variances (\eqn{\sigma^2}).}

\item{F}{Filter matrix
used to select the observed variables.}

\item{I}{Identity matrix.}

\item{SigmaMatrix}{Logical.
If \code{TRUE},
returns
the model-implied variance-covariance matrix
(\eqn{\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)}).
If \code{FALSE},
returns the
\eqn{\mathbf{S}}
matrix.}
}
\value{
Returns
the model-implied variance-covariance matrix
(\eqn{\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)})
or the
\eqn{\mathbf{S}}
matrix
derived from the
\eqn{\mathbf{A}}
matrix and
\eqn{\sigma^2}
vector.
}
\description{
Model-implied variance-covariance matrix
(\eqn{\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)})
or \eqn{\mathbf{S}} Matrix
using the Reticular Action Model notation.
}
\details{
Derives the
model-implied variance-covariance matrix
(\eqn{\boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right)})
or the
\eqn{\mathbf{S}}
matrix
from the
\eqn{\mathbf{A}}
matrix
and sigma squared
(\eqn{\sigma^2})
vector (variances).
\strong{Note that the first element in the matrix should be an exogenous variable.}
}
\examples{
A <- matrix(
  data = c(
    0, 0.26^(1 / 2), 0,
    0, 0, 0.26^(1 / 2),
    0, 0, 0
  ),
  ncol = 3
)
sigma2 <- c(15^2, 15^2, 15^2)
F <- I <- diag(3)
# Returns the model-implied variance-covariance matrix
ram_S(A = A, sigma2 = sigma2, F = F, I = I, SigmaMatrix = TRUE)
# Returns the model-implied S matrix
ram_S(A = A, sigma2 = sigma2, F = F, I = I, SigmaMatrix = FALSE)
}
\references{
McArdle, J. J. (2013).
The development of the RAM rules for latent variable structural equation modeling.
In A. Maydeu-Olivares & J. J. McArdle (Eds.),
\emph{Contemporary Psychometrics: A festschrift for Roderick P. McDonald} (pp. 225--273).
Lawrence Erlbaum Associates.

McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology, 37}(2), 234--251.
}
\seealso{
Other SEM notation functions: 
\code{\link{ram_M}()},
\code{\link{ram_Sigmatheta}()},
\code{\link{ram_mu}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{SEM notation functions}
\keyword{matrix}
\keyword{ram}
