% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{ram_Sigmatheta}
\alias{ram_Sigmatheta}
\title{Reticular Action Model - Model-Implied Variance Covariance Matrix
\eqn{\left( \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right) \right)}}
\usage{
ram_Sigmatheta(A, S, F, I, filter = TRUE)
}
\arguments{
\item{A}{Matrix.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{S}{Matrix.
Symmetric paths (double-headed arrows),
such as variances and covariances.}

\item{F}{Matrix.
Filter matrix
used to select the observed variables.}

\item{I}{Matrix.
Identity matrix.}

\item{filter}{Logical.
If \code{TRUE}, the \code{F} matrix is used.
If \code{FALSE}, the \code{F} matrix is omitted and the argument \code{F} is ignored.
See \code{Value} and \code{Details} below for more information.}
}
\value{
Returns the model-implied variance-covariance matrix
\eqn{\left( \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right) \right)}
derived from the \eqn{\mathbf{A}},
\eqn{\mathbf{S}},
\eqn{\mathbf{F}}, and
\eqn{\mathbf{I}} matrices.
If \code{filter = TRUE}, the \code{F} matrix is used as usual.
\deqn{
    \mathbf{F}
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{-1}
    \mathbf{S}
    \left[
      \left(
        \mathbf{I} - \mathbf{A}
      \right)^{-1}
    \right]^{\prime}
    \mathbf{F}^{\prime} .
  }
If \code{filter = FALSE}, the \code{F} matrix is omitted.
\deqn{
  \left(
    \mathbf{I} - \mathbf{A}
  \right)^{-1}
  \mathbf{S}
  \left[
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{-1}
  \right]^{\prime}
} .
}
\description{
Derives the model-implied variance-covariance matrix
\eqn{\left( \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right) \right)}
using the Reticular Action Model notation.
}
\details{
The model-implied variance-covariance matrix
\eqn{\left( \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right) \right)}
as a function of Reticular Action Model matrices
is given by
\deqn{
    \boldsymbol{\Sigma}
    \left(
      \boldsymbol{\theta}
    \right)
    =
    \mathbf{F}
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{-1}
    \mathbf{S}
    \left[
      \left(
        \mathbf{I} - \mathbf{A}
      \right)^{-1}
    \right]^{\prime}
    \mathbf{F}^{\prime}
  }
where
\eqn{\mathbf{A}}
represents asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.
\eqn{\mathbf{S}}
represents symmetric paths (double-headed arrows),
such as variances and covariances.
\eqn{\mathbf{F}}
represents the filter matrix
used to select the observed variables.
\eqn{\mathbf{I}}
represents the identity matrix.

When all the variables
in the \eqn{\mathbf{A}} and \eqn{\mathbf{S}} matrices
are observed,
\deqn{
    \boldsymbol{\Sigma}
    \left(
      \boldsymbol{\theta}
    \right)
    =
    \mathbf{F}
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{-1}
    \mathbf{S}
    \left[
      \left(
        \mathbf{I} - \mathbf{A}
      \right)^{-1}
    \right]^{\prime}
    \mathbf{F}^{\prime}
    =
    \left(
      \mathbf{I} - \mathbf{A}
    \right)^{-1}
    \mathbf{S}
    \left[
      \left(
        \mathbf{I} - \mathbf{A}
      \right)^{-1}
    \right]^{\prime} , \\
    \text{ when }
    \mathbf{F} = \mathbf{I} .
  }
}
\examples{
#################################################
# Simple mediation model with observed variables.
#################################################
# parameters
alpha <- 0.5
beta <- 0.4
tauprime <- 0.25
sigma2X <- 1.00
sigma2epsilonM <- 0.75
sigma2epsilonY <- 0.6775
# A and S matrices only contain observed variables X, M, and Y
A <- matrix(data = 0, nrow = 3, ncol = 3)
A[2, 1] <- alpha
A[3, 1] <- beta
A[3, 2] <- tauprime
S <- diag(c(sigma2X, sigma2epsilonM, sigma2epsilonY))
F <- I <- diag(3)
ram_Sigmatheta(A = A, S = S, F = F, I = I)
# Equivalent
ram_Sigmatheta(A = A, S = S, I = I, filter = FALSE)
# A and S matrices also contain latent variables sigma2epsilonM and sigma2epsilonY
A <- matrix(data = 0, nrow = 5, ncol = 5)
A[2, 1] <- alpha
A[2, 4] <- 1
A[3, 1] <- beta
A[3, 2] <- tauprime
A[3, 5] <- 1
S <- diag(c(sigma2X, 0, 0, sigma2epsilonM, sigma2epsilonY))
I <- diag(nrow(A))
F <- matrix(data = 0, nrow = 3, ncol = 5)
F[1, 1] <- 1
F[2, 2] <- 1
F[3, 3] <- 1
# In this case since F is not equal to I, we CANNOT use filter = FALSE
ram_Sigmatheta(A = A, S = S, F = F, I = I)
#######################
# One-factor CFA model.
#######################
A <- matrix(data = 0, ncol = 6, nrow = 6)
for (i in 2:6) {
  A[i, 1] <- 0.5
}
S <- diag(c(1, 0.75, 0.75, 0.75, 0.75, 0.75))
F <- I <- diag(nrow(A))
F <- diag(nrow(A) - 1)
F <- cbind(0, F)
ram_Sigmatheta(
  A = A,
  S = S,
  F = F,
  I = I
)
}
\references{
McArdle, J. J. (2013).
The development of the RAM rules for latent variable structural equation modeling.
In A. Maydeu-Olivares & J. J. McArdle (Eds.),
\emph{Contemporary Psychometrics: A festschrift for Roderick P. McDonald} (pp. 225--273).
Lawrence Erlbaum Associates.

McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology}, \emph{37} (2), 234--251.
}
\seealso{
Other SEM notation functions: 
\code{\link{ram_M}()},
\code{\link{ram_S}()},
\code{\link{ram_e}()},
\code{\link{ram_mutheta}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{SEM notation functions}
\keyword{matrix}
\keyword{ram}
