% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{rammutheta}
\alias{rammutheta}
\title{Reticular Action Model - Model-Implied Mean Vector
\eqn{\boldsymbol{\mu} \left( \boldsymbol{\theta} \right)}}
\usage{
rammutheta(M, A, filter)
}
\arguments{
\item{M}{\verb{p x 1} numeric vector \eqn{\mathbf{M}_{p \times 1}}.
Mean structure. Vector of means and intercepts.}

\item{A}{\verb{p x p} numeric matrix
\eqn{\mathbf{A}_{p \times p}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{filter}{\verb{k x p} numeric matrix
\eqn{\mathbf{F}_{k \times p}}.
Filter matrix used to select variables.}
}
\value{
Returns the model-implied mean vector
\eqn{ \boldsymbol{\mu} \left( \boldsymbol{\theta} \right)}
derived from the \eqn{\mathbf{A}}, \eqn{\mathbf{F}}, \eqn{\mathbf{I}},
matrices and \eqn{\mathbf{M}} vector.
}
\description{
Derives the model-implied mean vector
\eqn{\boldsymbol{\mu} \left( \boldsymbol{\theta} \right)}
using the Reticular Action Model (RAM) notation.
}
\details{
The model-implied mean vector
\eqn{\boldsymbol{\mu} \left( \boldsymbol{\theta} \right)}
as a function of Reticular Action Model (RAM) matrices
is given by

\deqn{
    \boldsymbol{\mu} \left( \boldsymbol{\theta} \right)
    =
    \mathbf{F} \left( \mathbf{I} - \mathbf{A} \right)^{-1} \mathbf{M}
  }

where
\itemize{
\item \eqn{\mathbf{A}_{p \times p}} represents asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings,
\item \eqn{\mathbf{I}_{p \times p}} represents an identity matrix,
\item \eqn{\mathbf{F}_{k \times p}} represents the filter matrix
used to select the observed variables,
\item \eqn{\mathbf{M}_{p \times 1}} represents the mean structure,
that is, a vector of means and intercepts,
\item \eqn{k} number of observed variables,
\item \eqn{q} number of latent variables, and
\item \eqn{p} number of observed and latent variables, that is \eqn{k + q} .
}
}
\examples{
#-----------------------------------------------------------------------------
# Simple mediation model with observed variables------------------------------
# The variables in the model are X, M, and Y.
# All three are observed variables.
#-----------------------------------------------------------------------------
alpha <- 0.3385926
beta <- 0.4510391
tauprime <- 0.2076475
muX <- 70.18000
deltaM <- -20.70243
deltaY <- -12.71288
A <- matrix(data = 0, nrow = 3, ncol = 3)
A[2, 1] <- alpha
A[3, 1] <- beta
A[3, 2] <- tauprime
filter <- diag(3)
M <- c(X = muX, M = deltaM, Y = deltaY)
colnames(A) <- c("X", "M", "Y")
rownames(A) <- c("X", "M", "Y")
colnames(filter) <- c("X", "M", "Y")
rownames(filter) <- c("X", "M", "Y")
rammutheta(M = M, A = A, filter = filter)

#-----------------------------------------------------------------------------
# Simple mediation model with observed variables------------------------------
# and latent variable error terms---------------------------------------------
# The same mediation model can be parameterized by explicitly
# including the error terms epsilonM and epsilonY in the matrices.
# The variables in the model are X, M, Y, epsilonM, and epsilonY.
# X, M, and Y are observed variables.
# epsilonM, and epsilonY are latent variables.
#-----------------------------------------------------------------------------
A <- matrix(data = 0, nrow = 5, ncol = 5)
A[2, 1] <- alpha
A[2, 4] <- 1
A[3, 1] <- beta
A[3, 2] <- tauprime
A[3, 5] <- 1
filter <- matrix(data = 0, nrow = 3, ncol = 5)
filter[1, 1] <- 1
filter[2, 2] <- 1
filter[3, 3] <- 1
M <- c(X = muX, M = deltaM, Y = deltaY, epsilonM = 0, epsilonY = 0)
colnames(A) <- c("X", "M", "Y", "epsilonM", "epsilonY")
rownames(A) <- c("X", "M", "Y", "epsilonM", "epsilonY")
colnames(filter) <- c("X", "M", "Y", "epsilonM", "epsilonY")
rownames(filter) <- c("X", "M", "Y")
rammutheta(M = M, A = A, filter = filter)

# One-factor CFA model--------------------------------------------------------
# The variables in the model are eta, y1, y2, y3, y4, y5.
# eta is a latent variable. y1 to y5 are observed variables.
#-----------------------------------------------------------------------------
A <- matrix(data = 0, ncol = 6, nrow = 6)
loadings <- 0.5 # tau equivalence
for (i in 2:6) {
  A[i, 1] <- 0.5
}
filter <- diag(nrow(A) - 1)
filter <- cbind(0, filter)
M <- c(eta = 0, y1 = 0, y2 = 0, y3 = 0, y4 = 0, y5 = 0)
colnames(A) <- c("eta", "y1", "y2", "y3", "y4", "y5")
rownames(A) <- c("eta", "y1", "y2", "y3", "y4", "y5")
colnames(filter) <- c("eta", "y1", "y2", "y3", "y4", "y5")
rownames(filter) <- c("y1", "y2", "y3", "y4", "y5")
rammutheta(M = M, A = A, filter = filter)
}
\references{
McArdle, J. J. (2013).
The development of the RAM rules for latent variable structural equation modeling.
In A. Maydeu-Olivares & J. J. McArdle (Eds.),
\emph{Contemporary Psychometrics: A festschrift for Roderick P. McDonald} (pp. 225--273).
Lawrence Erlbaum Associates.

McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology}, \emph{37} (2), 234--251.
}
\seealso{
Other SEM notation functions: 
\code{\link{ramM}()},
\code{\link{ramSigmatheta}()},
\code{\link{ramres}()},
\code{\link{ramsigma2}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{SEM notation functions}
\keyword{matrix}
\keyword{ram}
