% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{ramsigma2}
\alias{ramsigma2}
\title{Reticular Action Model - The \eqn{\mathbf{S}} Matrix from \eqn{\sigma^2}}
\usage{
ramsigma2(sigma2, A, start = TRUE)
}
\arguments{
\item{sigma2}{Numeric vector.
Vector of variances \eqn{\sigma^2}.}

\item{A}{\verb{p x p} numeric matrix
\eqn{\mathbf{A}_{p \times p}}.
Asymmetric paths (single-headed arrows),
such as regression coefficients and factor loadings.}

\item{start}{Logical.
If \code{TRUE}, an exogenous variable is positioned as the first element in the matrices.
If \code{FALSE}, an exogenous variable is positioned as the last element in the matrices.}
}
\value{
Returns the  \eqn{ \mathbf{S} } matrix.
}
\description{
Derives the \eqn{\mathbf{S}} matrix
using the Reticular Action Model (RAM) notation
from variable variances \eqn{\sigma^2}.
The off-diagonal elements of the \eqn{\mathbf{S}} matrix are assumed to be zeroes.
}
\details{
The \eqn{\mathbf{S}} matrix is derived using the \eqn{\mathbf{A}} matrix
and sigma squared \eqn{\left( \sigma^2 \right)} vector (variances).
\strong{Note that the first or last (see \code{start} argument) element
in the \code{A} and \code{S} matrices should be an exogenous variable.}
}
\examples{
#-----------------------------------------------------------------------------
# Simple mediation model with observed variables------------------------------
# The variables in the model are X, M, and Y.
# All three are observed variables.
#-----------------------------------------------------------------------------
alpha <- 0.3385926
beta <- 0.4510391
tauprime <- 0.2076475
sigma2X <- 1.2934694
sigma2M <- 1.0779592
sigma2Y <- 1.2881633
A <- matrix(data = 0, nrow = 3, ncol = 3)
A[2, 1] <- alpha
A[3, 1] <- beta
A[3, 2] <- tauprime
sigma2 <- c(X = sigma2X, M = sigma2M, Y = sigma2Y)
colnames(A) <- c("X", "M", "Y")
rownames(A) <- c("X", "M", "Y")
ramsigma2(sigma2 = sigma2, A = A)

#-----------------------------------------------------------------------------
# Simple mediation model with observed variables------------------------------
# and latent variable error terms---------------------------------------------
# The same mediation model can be parameterized by explicitly
# including the error terms epsilonM and epsilonY in the matrices.
# The variables in the model are X, M, Y, epsilonM, and epsilonY.
# X, M, and Y are observed variables.
# epsilonM, and epsilonY are latent variables.
#-----------------------------------------------------------------------------
A <- matrix(data = 0, nrow = 5, ncol = 5)
A[2, 1] <- alpha
A[2, 4] <- 1
A[3, 1] <- beta
A[3, 2] <- tauprime
A[3, 5] <- 1
sigma2 <- c(X = sigma2X, M = sigma2M, Y = sigma2Y)
colnames(A) <- c("X", "M", "Y", "epsilonM", "epsilonY")
rownames(A) <- c("X", "M", "Y", "epsilonM", "epsilonY")
ramsigma2(sigma2 = sigma2, A = A)

# One-factor CFA model--------------------------------------------------------
# The variables in the model are eta, y1, y2, y3, y4, y5.
# eta is a latent variable. y1 to y5 are observed variables.
#-----------------------------------------------------------------------------
A <- matrix(data = 0, ncol = 6, nrow = 6)
loadings <- 0.5 # tau equivalence
for (i in 2:6) {
  A[i, 1] <- 0.5
}
sigma2 <- c(eta = 1, y1 = 0.75, y2 = 0.75, y3 = 0.75, y4 = 0.75, y5 = 0.75)
colnames(A) <- c("eta", "y1", "y2", "y3", "y4", "y5")
rownames(A) <- c("eta", "y1", "y2", "y3", "y4", "y5")
ramsigma2(sigma2 = sigma2, A = A)

# One-factor CFA model--------------------------------------------------------
# The variables in the model are eta, y1, y2, y3, y4, y5.
# eta is a latent variable. y1 to y5 are observed variables.
#-----------------------------------------------------------------------------
A <- matrix(data = 0, ncol = 6, nrow = 6)
loadings <- 0.5 # tau equivalence
for (i in 1:5) {
  A[i, 6] <- 0.5
}
sigma2 <- c(y1 = 0.75, y2 = 0.75, y3 = 0.75, y4 = 0.75, y5 = 0.75, eta = 1)
colnames(A) <- c("y1", "y2", "y3", "y4", "y5", "eta")
rownames(A) <- c("y1", "y2", "y3", "y4", "y5", "eta")
ramsigma2(sigma2 = sigma2, A = A, start = FALSE)
}
\references{
McArdle, J. J. (2013).
The development of the RAM rules for latent variable structural equation modeling.
In A. Maydeu-Olivares & J. J. McArdle (Eds.),
\emph{Contemporary Psychometrics: A festschrift for Roderick P. McDonald} (pp. 225--273).
Lawrence Erlbaum Associates.

McArdle, J. J., & McDonald, R. P. (1984).
Some algebraic properties of the Reticular Action Model for moment structures.
\emph{British Journal of Mathematical and Statistical Psychology}, \emph{37} (2), 234--251.
}
\seealso{
Other SEM notation functions: 
\code{\link{ramM}()},
\code{\link{ramSigmatheta}()},
\code{\link{rammutheta}()},
\code{\link{ramres}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{SEM notation functions}
\keyword{matrix}
\keyword{ram}
